refactor(cursor): Rename `pg-cursor` package to `gaussdb-cursor`
