+++ packages/gaussdb-cursor/esm/index.mjs
@@ -1,4 +1,4 @@
-// ESM wrapper for pg-cursor
+// ESM wrapper for gaussdb-cursor
 import Cursor from '../index.js'
 
 // Export as default only to match CJS module
