+++ packages/gaussdb-cursor/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "pg-cursor",
+  "name": "gaussdb-cursor",
   "version": "2.15.0",
   "description": "Query cursor extension for GaussDB",
   "main": "index.js",
@@ -19,7 +19,7 @@
   "repository": {
     "type": "git",
     "url": "git@github.com:HuaweiCloudDeveloper/gaussdb-node.git",
-    "directory": "packages/pg-cursor"
+    "directory": "packages/gaussdb-cursor"
   },
   "author": "Brian M. Carlson",
   "license": "MIT",
