+++ packages/gaussdb-cursor/test/close.js
