+++ packages/gaussdb-cursor/test/error-handling.js
