+++ packages/gaussdb-cursor/test/no-data-handling.js
