+++ packages/gaussdb-cursor/test/promises.js
