+++ packages/gaussdb-cursor/test/query-config.js
