+++ packages/pg-esm-test/package.json
@@ -15,7 +15,7 @@
   "devDependencies": {
     "gaussdb": "^8.16.0",
     "gaussdb-cloudflare": "^1.2.5",
-    "pg-cursor": "^2.15.0",
+    "gaussdb-cursor": "^2.15.0",
     "pg-native": "^3.5.0",
     "pg-pool": "^3.10.0",
     "pg-protocol": "^1.10.0",
