+++ packages/pg-esm-test/pg-cursor.test.js
@@ -1,8 +1,8 @@
 import assert from 'node:assert'
 import { describe, it } from 'node:test'
-import Cursor from 'pg-cursor'
+import Cursor from 'gaussdb-cursor'
 
-describe('pg-cursor', () => {
+describe('gaussdb-cursor', () => {
   it('should export Cursor constructor as default', () => {
     assert.ok(new Cursor())
   })
