+++ packages/pg-pool/test/submittable.js
@@ -1,5 +1,5 @@
 'use strict'
-const Cursor = require('pg-cursor')
+const Cursor = require('gaussdb-cursor')
 const expect = require('expect.js')
 const describe = require('mocha').describe
 const it = require('mocha').it
